+++ NEWS.md
@@ -1,6 +1,7 @@
 # dbx 0.2.8 (unreleased)
 
 - Added support for `SQL` with `returning` option
+- Fixed time zones for RPostgres 1.3.0+ and RMariaDB 1.1.0+
 
 # dbx 0.2.7 (2020-09-07)